backend/post/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
python-dotenv==0.13.0
Flask-SQLAlchemy==2.4.4
redis==3.5.3
boto3==1.14.60
SQLAlchemy==1.3.19
psycopg2==2.8.6
kafka-python==2.0.1
requests==2.18.4
PyJWT==1.7.1
dnspython